Extract classes to ApiObjects.
