+++ lib/active_model/serializer/adapter/json_api/api_objects.rb
@@ -1,0 +1,16 @@
+require 'active_model/serializer/adapter/json_api/api_objects/link'
+require 'active_model/serializer/adapter/json_api/api_objects/resource'
+require 'active_model/serializer/adapter/json_api/api_objects/resource_identifier'
+require 'active_model/serializer/adapter/json_api/api_objects/relationship'
+require 'active_model/serializer/adapter/json_api/api_objects/json_api_object'
+
+module ActiveModel
+  class Serializer
+    module Adapter
+      class JsonApi
+        module ApiObjects
+        end
+      end
+    end
+  end
+end
