+++ lib/active_model/serializer/adapter/json_api/api_objects/resource.rb
@@ -1,0 +1,20 @@
+module ActiveModel
+  class Serializer
+    module Adapter
+      class JsonApi
+        module ApiObjects
+          Resource = Struct.new(:identifier, :attributes, :relationships, :links) do
+            def to_h
+              hash = identifier.to_h
+              hash[:attributes] = attributes if attributes.any?
+              hash[:relationships] = Hash[relationships.map { |k, v| [k, v.to_h] }] if relationships.any?
+              hash[:links] = links if links.any?
+
+              hash
+            end
+          end
+        end
+      end
+    end
+  end
+end
